+++ test/fixtures/privileges.sql
@@ -82,6 +82,8 @@ GRANT ALL ON TABLE
     , "UnitTest"
     , json_arr
     , jsonb_test
+    , authors_books_number
+    , authors_have_book_in_decade
 TO postgrest_test_anonymous;
 
 GRANT INSERT ON TABLE insertonly TO postgrest_test_anonymous;